Reorginzation of os_remote* files and lio_inspect fix to handle missing RIDs
This should be split into 2 patches.  The first just does a reorganization of the remote object service files
in anticipation of doing a full implementation.  Which I'll be working on next after taking a side project of tweaking
the inspection process.

The 2nd task is more substantive. It fixes an error when running inspect and the RID used in the eXnode doesn't appear in the RS.
This is classified as a soft error and is not reported by default.  Assuming that's the only issue with the allocation.
If the user adds the INSPECT_SOFT_ERROR_FAIL flag to the inspect mode then this case is treated as a hard error and reported.