Fixed a bug resulting in hconnections being destroyed before being closed.
The host portal was being closed or compacted without waiting for connections that
were in the process of shutting down.  This led to odd segfaults.  It was hard to
detect using valgrind as well since the APR memory pool was properly cleaning up
in most instances.  Even though the original thread was still holding references to variables
in the freed mpool. The memory pool would then be reused by another thread keeping
it in scope and used by 2 threads.