Added support for specifying include paths to search for included files
You can now add the directive:

%include_path /my/include/path

with /my/include/path replaced with the path of your choice.  The directive can be
added multiple times. The paths are searched in the order they added.

This means you should only have to add a few of these directives and then use relative
path names for all the other include files.