Refactored MQ incorpotation and base support for extra OS abort calls
This version has the refacted MQ support which resulted in the addition of a new "ongoing"
object added to the lio_config ess grab bag.  It doesn't have an explicit lio_config_t entry.
Instead it's just in the ess grab bag that any routine can get access to.

I've also aded 2 additional abort calls to the OS
	abort_remove_regex_object(), abort_regex_object_set_multiple_attrs()

These weren't needed until the remote OS was created cause a simple ^C would do the same thing.
But with the advent of the remote OS this is no longer the case.  A ^C on the client just kills the client.
It would not stop the server from continuing the remove or set operation.
Only the os_simple implementation actually supports the new routines.  A subsequent commit in the coming days
will add the support to the remote OS client/server implmentations.