Added ability to abort os_remove_regex_object() and os_regex_set_multiple_attrs().
This wasn't really an issue with os_file but with the introdution of the os_remote* support this
created a dangerous edge case.  The problem is if the os_remote_client calls either of the above routines
and it's received by the client there was no way to stop the task from completing without killing the os_remote_server
which would not be possible in a distributed environment.  2 new routines have been added to programmatically abort
both of these commands.  Additionally they also require the client to constantly send "I'm Alive" packets.  As long
as those keep coming the the task will continue running on the server.  If they stop the server automatically aborts the task.
This way if the client is ^C-ed a few seconds later the server will abort as well.  I've been using "I'm Alive" packets being
sent at 1 sec intervals and the abort is 4 secs.