Added command specific environment parsing
Previously only the LIO_OPTIONS envirment variable was checked for additional arguments.
To make handling the various configs I've added a command specific env check 1st with the global
option as the fallback.  The environment variable look at is based on the executable name stored in
argv[0]. The executable basename is used to construct the env variable.  If the name is lio_XXXX then
the env checked is LIO_OPTIONS_XXXX otherwise the whole basename is used -- LIO_OPTIONS_basename.