Initial retooling of the cache.  **NOTE: cache_lru has not been updated and will not work!**
This commit eliminates the atomic_int declarations for the access_pending[] fields on each page.  They are now normal integers.
Additionally only the segment size, segment cache stats, and segment flushing stack are protected by the segment lock.  Everything
else is protected by the global cache lock.

Before if pages were available for access they were flagged 1st and then the data was R/W to them based on the mode and then released.
This has the benefit of allowing memory copies to be done outside the lock but also resulted in alot of book keeping negating the benefit.
Now if pages are available they are immediately read/written inside the cache lock.  Pages are only flagged if data needs to be transferred
from a depot.  Flagged pages are then treated the same as before.  This also eliminated several hoops on releasing pages simplifying that logic as well.

The partial page support has the potential for duplicate pages (both a partial page and a "normal" page) to exist at the same time with different contents.
You acheive eventual consistency but this is not ideal.  The logic dealing with ppages is also pretty heavyweight and will be simplified in the next pass.