Minimally populate file path
Fixed a few bugs but the real purpose of the patch is to add the ability
to minimally populate a file path. The way FUSE uses the driver is it
will initially recurse down through the path collecting attributes as it
goes but when those attributes time out it doesn't go back and refresh
them.  Instead it will just keep probing the specific file attribute.
So for example on a write it will check a few posix ACL's on each write.
Since the prefix attributes have expired all the OS attribute cache will
have been purged as well so an attempt to add the open file attributes
will fail because the prefix is not stored in the cache tree.  This
patch minimally populates the prefix so the terminal can be stored.

This has not been checked to work against harlinks, symbolic links, and
rename operations.  That will be done next.