Initial support for Ubuntu/Debian & deb packaging
Added init-source.sh script so it is possible to initialize a fresh
source tree for building packages in docker without first building
everything on the host OS

Additionally added czmq build for distros lacking that package.

These are the current debian/ubuntu releases that we're able to use. To
get support for other (older) releases, it appears we'll need to pack
along our own libzmq3. At some point, we should figure out if that's
worthwhile