Build repositories and test their functionality
This commit adds two new entry scripts - update-repo.sh and
test-repo.sh.

The update-repo.sh takes the packages built by package.sh and stored in
lstore-release/packages and produces/updates a repository. Currently
only RPM is supported, but Matt H. is working on the DEB half.

The test-repo.sh loads a stock container, adds the repository from the
previous step to the package management system, and attempts to install
the packages and run some tests to make sure all the symbols resolve
properly.