Skeleton for common CMake framework
Having a global CMake configuration will greatly simplify building
LStore variants (e.g. static analysis, address sanitation). Start the
work on that branch with a basic CMakeLists.txt that builds the
externals, and an additional test case in the Jenkinsfile to make sure
both the old and new ways work.