Replace list of preprocessor constants with enums
A lot of lists are defined as a series of preprocessor #defines. This is
harmful because of a lack of type-safety and the inability for
debugging to be able to backtrace the meaning of arbitrary integer
constants.

Instead, replace preprocessor #defines with enum types, which can then
replace "int" throughout the code