gridftp: Add basic stat support
Add stat support and a basic test for it. The test will only work on the
CMS LServer (since it depends on a specific file existing).

The bigger problem is that I got very frustrated with LStore directory
listing/stat API and punted for later to finish the support for listing
directory contents. See binding/gridftp/stat.c for the additional note.