gridftp: Add checksumming, refactor xfer paths
In the process of adding checksumming, it became apparent the xfer paths
had become too complicated. Recombine the current xfer paths at the gfs
layer so they can then be refactored back apart. Ideally the gfs layer
performs all the munging of the callback interface w/gridftp so the user
layer can just handle the reading and writing of the files