Fix ZMQ bugs resulting from dropping CZMQ and add DNS TCP references. We use an extension to the ZMQ_ROUTER protocol to support sending and receiving packets through multiple hosts.  When one binds or connects using ZMQ the address provided just establishes the physical link but does NOT provide a protocol connection.  The protocol requires an additional name which is stored as the ZMQ_IDENTITY for each socket and is set by the application.
Before the patch the address and protocol name were the same for bind
sockets and "address:random_number" for connect sockets.  This meant
that if the server used "tcp://localhost:6713" for the address and proto
id the client had to do the same.  So if the client specified
"tcp://127.0.0.1:6713" it would be able to make the TCP connection but
the ZMQ protocol would fail to send packets because the strings weren't
the same for both client and server.

As part of this patch I've added the ability to specify both a TCP address
and a protocol ID. The new string has the format:

    zmq_id|tcp://host:port

So now the server could specify "SERVER|tcp://localhost:6713" and the
client can use "SERVER|tcp://127.0.0.1:6713" and connect at both the
link and protocol level.

A connect socket uses an automatically generated ID of
"hostname:random_number".  We could extend connect sockets to take a
protocol ID but it's not needed for us at this time.

With this change we can now use all the various methods of specifying the
physical link that ZMQ supports.  So these all work for the server bind:

    SERVER|tcp://*:6713          - Listen on all interfaces
    SERVER|tcp://eth0:6713       - Listen on the eth0 interface
    SERVER|tcp://localhost:6713
    SERVER|tcp://127.0.0.1:6713

ZMQ also supports IPv6 addresses.