Add QoS to mq_stream. The old behavior defaulted to disabling the background heartbeating to keep connections alive.  This was done to save the extra overhead of managing the hearbeats.  Data would then be flushed only when the buffer was full or the heartbeat timeout was reached which was typically 60s.  Such a long interval with no tangible response leads one to think the command is hung when it really isn't.
The change enables heartbeating for all stream operations that have path
recursion enabled.  Additionally opportunistic flushing occurs after no
data is sent in more than 1s.  This way a steady stream of data comes back
to the client.  It's opportunistic because it's triggered in the stream_write()
function.  The heartbeat timeout handles the no activity case.