The new default is to look at /etc/lio/default for where to load the config. If this file is missing it falls back to the old procedure. The file should contain a single line pointing to a remote config server or a local INI file with the config.  The formats are as follows:
    [lstore://][user@][MQ_NAME|]HOST[:port:][:cfg][:section]
        Get the config from a remote LServer

    [ini://]/path/to/ini_file
        Local INI config file

The normal "-c" option also supports passing a string using either of the
above formats along with a version that points to an alternative default
file which contains the single line mentioned above.  The default format is

    file:///path/to/default

To enable an LServer to act as a remote config server you need to add "host"
and "prefix" variables in the INI section being loaded. The default is to load "[lio]".
The host has the same format as other MQ hosts.  The "prefix" specifies where
to look for client configs to send.  The "cfg" name sent by the client gets
munged into "${prefix}/${cfg}.cfg".  If this file exists it's contents are
returned.  All included files are transparently resolved before sending.

The patch also sanitizies the object reference counting.  It now makes more
logical sense.